currency_weather_service/requirements.txt
requests
pystray
Pillow
plyer
